+++ packages/dsv-charts/src/charts/index.ts
@@ -1,5 +1,1 @@
-import { IArrayChartConfig } from './array';
-
-export type IConfig = IArrayChartConfig;
-
 export * from './array';
