+++ packages/dsv-charts/src/components/layouts/index.ts
@@ -1,3 +1,2 @@
 export * from './cartesian2';
-export * from './type';
 export * from './base';
