+++ packages/dsv-charts/src/index.ts
@@ -1,1 +1,4 @@
-export * from './old-charts';
+// export * from './old-charts';
+export * from './data-structure';
+export * from './charts';
+export { Dsv as default } from './core';
