--- packages/dsv-charts/src/theme/index.ts
@@ -1,2 +1,0 @@
-export * from './dark';
-export * from './light';
