--- packages/dsv-charts/src/theme/light.ts
@@ -1,23 +1,0 @@
-import { ITheme } from '@dsv-charts/typings';
-
-export const light: ITheme = {
-  colorScheme: [
-    '#93b7e3',
-    '#59c4e6',
-    '#edafda',
-    '#516b91',
-    '#a5e7f0',
-    '#cbb0e3',
-  ],
-  text: {
-    textColor: '#212121',
-  },
-  border: {
-    color: '#eeeeee',
-    width: 10,
-  },
-  arrow: {
-    color: '#212121',
-    width: 2,
-  },
-};
