--- packages/dsv-charts/src/utils/array.ts
@@ -1,9 +1,0 @@
-let uniqueKey = 0;
-
-export function createArrayItem(value: string | number) {
-  return {
-    key: `__array__key__${uniqueKey++}`,
-    value: value,
-    name: value.toString(),
-  };
-}
