--- packages/dsv-charts/src/utils/stack.ts
@@ -1,9 +1,0 @@
-let uniqueKey = 0;
-
-export function createStackItem(value: string | number) {
-  return {
-    key: `__stack__key__${uniqueKey++}`,
-    value: value,
-    name: value.toString(),
-  };
-}
