+++ packages/dsv-website/src/config/data-structure/array/fill.js
@@ -1,6 +1,6 @@
 const code = `const array = dsv.create({
-type: 'array',
-data: [10, 8, 6, 4, 2, 1, 3, 5, 7, 9]
+  type: 'array',
+  data: [10, 8, 6, 4, 2, 1, 3, 5, 7, 9]
 })
 
 array.fill(5, 5)
